pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "compspace"
version = "0.0.6"
description = "A matplotlib-based plotting library for visualizing composition spaces."
authors = [
    { name = "Felix Thelen", email = "felix.thelen@ruhr-uni-bochum.de" }
]
license = { file = "LICENSE" }
readme = "README.md"
keywords = ["materials science", "composition spaces", "chemistry", "matplotlib"]
requires-python = ">=3.9"
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    "numpy >= 1.26.0",
    "pandas >= 2.1.1",
    "matplotlib >= 3.1.0",
    "ffmpeg-python >= 0.2.0"
]

[metadata]
license_expression = "GPL-3.0-or-later"

[project.urls]
Repository = "https://github.com/felixthln/simplices"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
